move weather files into weather subfolder
